+++ bench/priv/repo/migrations/20260916000002_add_scriba_watermarks.exs
@@ -1,0 +1,7 @@
+defmodule ScribaBench.Repo.Migrations.AddScribaWatermarks do
+  use Ecto.Migration
+
+  # The upgrade path: this database was created at schema version 1.
+  def up, do: Scriba.Migrations.up(from: 1)
+  def down, do: Scriba.Migrations.down(to: 1)
+end